==> code/02-MergeData.ipynb <==
Import libraries

from matplotlib import pyplot as plt
import numpy as np
import pandas as pd

Read data from .csv files

edu_spending_DF = pd.read_csv("../data/oecd_education_spending.csv")
edu_private_DF = pd.read_csv("../data/oecd_private_spending_on_education.csv")
edu_public_DF = pd.read_csv("../data/oecd_public_spending_on_education.csv")
edu_tertiary_DF = pd.read_csv("../data/oecd_spending_on_tertiary_education.csv")

pisa_reading_DF = pd.read_csv("../data/oecd_pisa_reading.csv")
pisa_math_DF = pd.read_csv("../data/oecd_pisa_math.csv")
pisa_science_DF = pd.read_csv("../data/oecd_pisa_science.csv")

researchers_rd_DF = pd.read_csv("../data/wb_researchers_in_rd.csv")
rd_expenditure_DF = pd.read_csv("../data/wb_rd_expenditure.csv")

high_tech_exports_DF = pd.read_csv("../data/wb_high_technology_exports.csv")

population_DF = pd.read_csv("../data/wb_population_total.csv")

patent_res_DF = pd.read_csv("../data/wb_patent_applications_residents.csv")
patent_nres_DF = pd.read_csv("../data/wb_patent_applications_nonresidents.csv")

gini_DF = pd.read_csv("../data/wb_gini_index.csv")
gdp_percapita_DF = pd.read_csv("../data/wb_gdp_per_capita.csv")

Convert data frames into lists

edu_spending = edu_spending_DF.values
edu_private = edu_private_DF.values
edu_public = edu_public_DF.values
edu_tertiary = edu_tertiary_DF.values

pisa_reading = pisa_reading_DF.values
pisa_math = pisa_math_DF.values
pisa_science = pisa_science_DF.values

researchers_rd = researchers_rd_DF.values
rd_expenditure = rd_expenditure_DF.values

high_tech_exports = high_tech_exports_DF.values

population = population_DF.values
patent_res = patent_res_DF.values
patent_nres = patent_nres_DF.values

gini = gini_DF.values
gdp_percapita = gdp_percapita_DF.values

Education Spendings -- Primary to Non-tertiary (Latest)

edu_spending_pryntry = [[x[0], x[5], x[6]] for x in edu_spending if x[0] != "OAVG" and x[2] == "PRY_NTRY" and ~np.isnan(x[6])]

### get latest data
edu_spending_pryntry_latest = []
for i in range(len(edu_spending_pryntry) - 1):
    if edu_spending_pryntry[i][0] != edu_spending_pryntry[i + 1][0]:
        edu_spending_pryntry_latest.append(edu_spending_pryntry[i])
edu_spending_pryntry_latest.append(edu_spending_pryntry[-1])

Education Spendings -- Tertiary (Latest)

edu_spending_try = [[x[0], x[5], x[6]] for x in edu_spending if x[0] != "OAVG" and x[2] == "TRY" and ~np.isnan(x[6])]

### get latest data
edu_spending_try_latest = []
for i in range(len(edu_spending_try) - 1):
    if edu_spending_try[i][0] != edu_spending_try[i + 1][0]:
        edu_spending_try_latest.append(edu_spending_try[i])
edu_spending_try_latest.append(edu_spending_try[-1])

Create a data list

Add pryntry, try into list

data = []
header = ["name", "year", "pryntry", "try", "total"]

for i in range(41):
    name = edu_spending_pryntry_latest[i][0]
    year = edu_spending_pryntry_latest[i][1]
    
    data_pryntry = round(edu_spending_pryntry_latest[i][2], 2)
    data_try = round(edu_spending_try_latest[i][2], 2)
    
    data_total = round(data_pryntry + data_try, 2)
    data.append([name, year, data_pryntry, data_try, data_total])   

Create 2 additional lists for name and year

data_name = [];
data_year = [];

for i in range(41):
    data_name.append(data[i][0])
    data_year.append(data[i][1])

Add tertiary_public into list

header.append("try_public")

edu_tertiary_public = [[x[0], x[5], x[6]] for x in edu_tertiary if x[0] != "OAVG" and x[2] == "PUB" and ~np.isnan(x[6])]

for i in range(len(edu_tertiary_public)):
    if edu_tertiary_public[i][0] in data_name:
        index = data_name.index(edu_tertiary_public[i][0])
        if edu_tertiary_public[i][1] == data_year[index]:
            data[index].append(round(edu_tertiary_public[i][2], 2))

for i in range(len(edu_tertiary_public)):
    if edu_tertiary_public[i][0] in data_name:
        index = data_name.index(edu_tertiary_public[i][0])
        if len(data[index]) == 5 and edu_tertiary_public[i][1] == data_year[index] - 1:
            data[index].append(round(edu_tertiary_public[i][2], 2))    
            
for i in range(len(edu_tertiary_public)):
    if edu_tertiary_public[i][0] in data_name:
        index = data_name.index(edu_tertiary_public[i][0])
        if len(data[index]) == 5 and edu_tertiary_public[i][1] == data_year[index] + 1:
            data[index].append(round(edu_tertiary_public[i][2], 2))    
                    
for x in data:
    if len(x) == 5:
        x.append(None)

Add tertiary_private into list

header.append("try_private")

edu_tertiary_private = [[x[0], x[5], x[6]] for x in edu_tertiary if x[0] != "OAVG" and x[2] == "PRIV" and ~np.isnan(x[6])]

for i in range(len(edu_tertiary_private)):
    if edu_tertiary_private[i][0] in data_name:
        index = data_name.index(edu_tertiary_private[i][0])
        if edu_tertiary_private[i][1] == data_year[index]:
            data[index].append(round(edu_tertiary_private[i][2], 2))

for i in range(len(edu_tertiary_private)):
    if edu_tertiary_private[i][0] in data_name:
        index = data_name.index(edu_tertiary_private[i][0])
        if len(data[index]) == 6 and edu_tertiary_private[i][1] == data_year[index] - 1:
            data[index].append(round(edu_tertiary_private[i][2], 2))            
            
for i in range(len(edu_tertiary_private)):
    if edu_tertiary_private[i][0] in data_name:
        index = data_name.index(edu_tertiary_private[i][0])
        if len(data[index]) == 6 and edu_tertiary_private[i][1] == data_year[index] + 1:
            data[index].append(round(edu_tertiary_private[i][2], 2))      
            
for x in data:
    if len(x) == 6:
        x.append(None)

Convert try_public / try into try_public
Convert try_private / try into try_private

for x in data:
    if x[5] is not None and x[6] is not None:
        x[5] = round(x[5] * x[3] / 100, 2)
        x[6] = round(x[6] * x[3] / 100, 2)

Get latest PISA scores

pisa_read_total = [[x[0], x[5], x[6]] for x in pisa_reading if x[0] != "OAVG" and x[2] == "TOT" and ~np.isnan(x[6])]
pisa_math_total = [[x[0], x[5], x[6]] for x in pisa_math if x[0] != "OAVG" and x[2] == "TOT" and ~np.isnan(x[6])]
pisa_science_total = [[x[0], x[5], x[6]] for x in pisa_science if x[0] != "OAVG" and x[2] == "TOT" and ~np.isnan(x[6])]### get latest data

### get latest data

pisa_read_total_latest = []
for i in range(len(pisa_read_total) - 1):
    if pisa_read_total[i][0] != pisa_read_total[i + 1][0]:
        pisa_read_total_latest.append(pisa_read_total[i])
pisa_read_total_latest.append(pisa_read_total[-1])

pisa_math_total_latest = []
for i in range(len(pisa_math_total) - 1):
    if pisa_math_total[i][0] != pisa_math_total[i + 1][0]:
        pisa_math_total_latest.append(pisa_math_total[i])
pisa_math_total_latest.append(pisa_math_total[-1])

pisa_science_total_latest = []
for i in range(len(pisa_science_total) - 1):
    if pisa_science_total[i][0] != pisa_science_total[i + 1][0]:
        pisa_science_total_latest.append(pisa_science_total[i])
pisa_science_total_latest.append(pisa_science_total[-1])

Add PISA scores into list

header.append("pisa_read")
header.append("pisa_math")
header.append("pisa_science")

for i in range(44):
    name = pisa_read_total_latest[i][0]
    
    if name in data_name:
        index = data_name.index(name)
        
        data[index].append(pisa_read_total_latest[i][2])
        data[index].append(pisa_math_total_latest[i][2])
        data[index].append(pisa_science_total_latest[i][2])

for x in data:
    if len(x) == 7:
        x.append(None)
        x.append(None)
        x.append(None)

Add rd_expenditure into list

header.append("rd_exp")  

for x in rd_expenditure:
    if x[1] in data_name:
        index = data_name.index(x[1])
        year = data_year[index]
        
        if ~np.isnan(x[year-2018]):
            data[index].append(round(x[year-2018], 2))
        elif ~np.isnan(x[year-2017]):
            data[index].append(round(x[year-2017], 2))
        elif ~np.isnan(x[year-2019]):
            data[index].append(round(x[year-2019], 2))   
            
for x in data:
    if len(x) == 10:
        x.append(None)

Add researchers/population*1000000 into list

header.append("researchers_rd_permillionpeople") 

for i, x in enumerate(researchers_rd):
    if x[1] in data_name:
        index = data_name.index(x[1])
        year = data_year[index]
        
        if ~np.isnan(x[year-2018]):
            data[index].append(round(x[year-2018]/population[i][year-2018]*1000000, 2))
        elif ~np.isnan(x[year-2017]):
            data[index].append(round(x[year-2017]/population[i][year-2017]*1000000, 2))
        elif ~np.isnan(x[year-2019]):
            data[index].append(round(x[year-2019]/population[i][year-2019]*1000000, 2))    
        
for x in data:
    if len(x) == 11:
        x.append(None)

Add high_tech_exports into list

header.append("high_tech_exports")

for x in high_tech_exports:
    if x[1] in data_name:
        index = data_name.index(x[1])
        year = data_year[index]
        
        if ~np.isnan(x[year-2018]):
            data[index].append(round(x[year-2018], 2))
        elif ~np.isnan(x[year-2017]):
            data[index].append(round(x[year-2017], 2))
        elif ~np.isnan(x[year-2019]):
            data[index].append(round(x[year-2019], 2))              

for x in data:
    if len(x) == 12:
        x.append(None)

Add patent_res/population*100 into list

header.append("patent_res_per100people")    

for i, x in enumerate(patent_res):
    if x[1] in data_name:
        index = data_name.index(x[1])
        year = data_year[index]
        
        if ~np.isnan(x[year-2018]):
            data[index].append(round(x[year-2018]/population[i][year-2018]*1000, 2))
        elif ~np.isnan(x[year-2017]):
            data[index].append(round(x[year-2017]/population[i][year-2017]*1000, 2))
        elif ~np.isnan(x[year-2019]):
            data[index].append(round(x[year-2019]/population[i][year-2019]*1000, 2))    
            
for x in data:
    if len(x) == 13:
        x.append(None)

Add patent_nres/population*100 into list

header.append("patent_nres_per100people") 

for i, x in enumerate(patent_nres):
    if x[1] in data_name:
        index = data_name.index(x[1])
        year = data_year[index]
        
        if ~np.isnan(x[year-2018]):
            data[index].append(round(x[year-2018]/population[i][year-2018]*1000, 2))
        elif ~np.isnan(x[year-2017]):
            data[index].append(round(x[year-2017]/population[i][year-2017]*1000, 2))
        elif ~np.isnan(x[year-2019]):
            data[index].append(round(x[year-2019]/population[i][year-2019]*1000, 2))  
    
for x in data:
    if len(x) == 14:
        x.append(None)

Add gdp_percapita/1000 into list

header.append("gdp_percapita_dividedby1000") 

for x in gdp_percapita:
    if x[1] in data_name:
        index = data_name.index(x[1])
        year = data_year[index]
        
        if ~np.isnan(x[year-2018]):
            data[index].append(round(x[year-2018]/1000, 2))
        elif ~np.isnan(x[year-2017]):
            data[index].append(round(x[year-2017]/1000, 2))
        elif ~np.isnan(x[year-2019]):
            data[index].append(round(x[year-2019]/1000, 2)) 
        
for x in data:
    if len(x) == 15:
        x.append(None)

Add gini into list

header.append("gini_index")  

for x in gini:
    if x[1] in data_name:
        index = data_name.index(x[1])
        year = data_year[index]
        
        if ~np.isnan(x[year-2018]):
            data[index].append(round(x[year-2018], 2))
        elif ~np.isnan(x[year-2017]):
            data[index].append(round(x[year-2017], 2))
        elif ~np.isnan(x[year-2019]):
            data[index].append(round(x[year-2019], 2))
        
for x in data:
    if len(x) == 16:
        x.append(None)

Sort data in alphabetical order

data = sorted(data, key=lambda data: data[0])

Add full names

header.insert(1, "full_name")

allnames = population[:, 0].tolist()
allcodes = population[:, 1].tolist()

for x in data:
    index = allcodes.index(x[0])
    x.insert(1, allnames[index])

Convert list to matrix and then to data frame

dm = np.matrix(data)
df = pd.DataFrame(dm, index = np.arange(1, 42), columns = header)

Write data frame into a .csv file

df.to_csv("../data/alldata.csv")
