==> code/03-FilterMissingData.ipynb <==
Import libraries

from matplotlib import pyplot as plt
import numpy as np
import pandas as pd

Read .csv file

df = pd.read_csv("../data/alldata.csv", index_col = 0)

Convert data frame into a matrix and a list

dm = df.values
header = df.axes[1].tolist()
data = dm.tolist()

Get only rows with no missing data

data_full = []

for x in data:
    remove = False
    for cell in x:
        if cell != cell:
            remove = True
    if remove == False: 
        data_full.append(x)

Create data frame

data = data_full
dm = np.matrix(data)
df = pd.DataFrame(dm, columns = header, index = np.arange(1, 27))

Create .csv file

df.to_csv("../data/alldata_full.csv")
