==> code/05-MinimizeData.ipynb <==
Import libraries

import numpy as np
import pandas as pd

Read data

df = pd.read_csv("../data/alldata_full.csv", index_col = 0)

Throw try and total since try_public + try_private = try and pryntry + try = total

df = df[['name', 'full_name', 'year', 'pryntry', 'try_public', 'try_private', 'pisa_read', 'pisa_math', 'pisa_science', 'rd_exp', 'researchers_rd_permillionpeople', 'high_tech_exports', 'patent_res_per100people', 'patent_nres_per100people', 'gdp_percapita_dividedby1000', 'gini_index']]

Write data

df.to_csv("../data/minimizeddata_full.csv")
