==> code/07-Cluster.ipynb <==
Import Libraries

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

Read data

df = pd.read_csv("../data/minimizeddata_full.csv", index_col = 0)

Cluster

from sklearn.cluster.bicluster import SpectralCoclustering

model = SpectralCoclustering(n_clusters=2, random_state=0, )
model.fit(df[['gdp_percapita_dividedby1000', 'gini_index']])

Add clusters to data frame

df.insert(3, "cluster", pd.Series(model.row_labels_, index=df.index)  + 1)

Convert to .csv

df.to_csv("../data/minimizeddata_full_clustered.csv")
