==> code/08-LinearRegressionNew.ipynb <==
Import libraries

from matplotlib import pyplot as plt
from matplotlib import rc
import numpy as np
import pandas as pd
import statsmodels.formula.api as smf

Read data

df = pd.read_csv("../data/minimizeddata_full.csv", index_col = 0)

Regression on pisa_read

lm_pisaread = smf.ols(formula='pisa_read ~ pryntry', data=df).fit()
lm_pisaread.summary()

Regression on pisa_math

lm_pisamath = smf.ols(formula='pisa_math ~ pryntry', data=df).fit()
lm_pisamath.summary()

Regression on pisa_science

lm_pisascience = smf.ols(formula='pisa_science ~ pryntry', data=df).fit()
lm_pisascience.summary()

COMMENT:

Regressions on pisa_read, pisa_math, pisa_science are all non-significant.

Regressions on rd_exp

lm_rdexp_unrestricted = smf.ols(formula='rd_exp ~ pisa_read + pisa_math + pisa_science + pryntry + try_public + try_private', data=df).fit()
lm_rdexp_unrestricted.summary()

lm_rdexp_restricted = smf.ols(formula='rd_exp ~ pisa_math + try_public', data=df).fit()
lm_rdexp_unrestricted.compare_f_test(lm_rdexp_restricted)

lm_rdexp = lm_rdexp_restricted
lm_rdexp.summary()

COMMENT:

When I perform a f-test on unrestricted and restricted regressions on rd_exp, I see that all random variables except pisa_math and try_public are jointly non-significant.
Therefore, I took the restricted regression.

Regressions on researchers_rd

lm_researchersinrd_unrestricted = smf.ols(formula='researchers_rd_permillionpeople ~ pisa_read + pisa_math + pisa_science + pryntry + try_public + try_private', data=df).fit()
lm_researchersinrd_unrestricted.summary()

lm_researchersinrd_restricted = smf.ols(formula='researchers_rd_permillionpeople ~ pryntry', data=df).fit()
lm_researchersinrd_unrestricted.compare_f_test(lm_researchersinrd_restricted)

lm_researchersinrd = lm_researchersinrd_restricted
lm_researchersinrd.summary()

COMMENT:

When I perform a f-test on unrestricted and restricted regressions on researchers_rd, I see that all random variables except try_public are jointly non-significant. Therefore, I took the restricted regression.

Regressions on high_tech_exports

lm_hightechexports_unrestricted = smf.ols(formula='high_tech_exports ~ pisa_read + pisa_math + pisa_science + pryntry + try_public + try_private + rd_exp + researchers_rd_permillionpeople', data=df).fit()
lm_hightechexports_unrestricted.summary()

COMMENT:

Regression on high technology exports are jointly non significant.

Regressions on patent_res_per100people

lm_residentpatent_unrestricted = smf.ols(formula='patent_res_per100people ~ pisa_read + pisa_math + pisa_science + pryntry + try_public + try_private + rd_exp + researchers_rd_permillionpeople', data=df).fit()
lm_residentpatent_unrestricted.summary()

lm_residentpatent_restricted = smf.ols(formula='patent_res_per100people ~ try_private + rd_exp', data=df).fit()
lm_residentpatent_unrestricted.compare_f_test(lm_residentpatent_restricted)

lm_residentpatent = lm_residentpatent_restricted
lm_residentpatent.summary()

COMMENT:

When I perform a f-test on unrestricted and restricted regressions on patent_res_per100people, I see that all random variables except try_private and rd_exp are jointly non-significant. Therefore, I took the restricted regression.

Regressions on patent_nres_per100people

lm_nonresidentpatent_unrestricted = smf.ols(formula='patent_nres_per100people ~ pisa_read + pisa_math + pisa_science + pryntry + try_public + try_private + rd_exp + researchers_rd_permillionpeople', data=df).fit()
lm_nonresidentpatent_unrestricted.summary()

lm_nonresidentpatent_restricted = smf.ols(formula='patent_nres_per100people ~ try_private + rd_exp', data=df).fit()
lm_nonresidentpatent_unrestricted.compare_f_test(lm_nonresidentpatent_restricted)

lm_nonresidentpatent = lm_nonresidentpatent_restricted
lm_nonresidentpatent.summary()

COMMENT:

When I perform a f-test on unrestricted and restricted regressions on patent_nres_per100people, I see that all random variables except try_private and rd_exp are jointly non-significant. Therefore, I took the restricted regression.

Regressions on gdp_percapita_dividedby1000

lm_gdppercapita_unrestricted = smf.ols(formula='gdp_percapita_dividedby1000 ~ pisa_read + pisa_math + pisa_science + pryntry + try_public + try_private + rd_exp + researchers_rd_permillionpeople + high_tech_exports + patent_res_per100people + patent_nres_per100people', data=df).fit()
lm_gdppercapita_unrestricted.summary()

lm_gdppercapita_restricted1 = smf.ols(formula='gdp_percapita_dividedby1000 ~ pisa_math + pryntry + try_public + try_private + patent_nres_per100people', data=df).fit()
lm_gdppercapita_unrestricted.compare_f_test(lm_gdppercapita_restricted1)

lm_gdppercapita_restricted1.summary()

lm_gdppercapita_restricted2 = smf.ols(formula='gdp_percapita_dividedby1000 ~ pryntry + try_public + try_private + patent_nres_per100people', data=df).fit()
lm_gdppercapita_restricted1.compare_f_test(lm_gdppercapita_restricted2)

lm_gdppercapita_restricted2.summary()

lm_gdppercapita_restricted3 = smf.ols(formula='gdp_percapita_dividedby1000 ~ pryntry + try_private + patent_nres_per100people', data=df).fit()
lm_gdppercapita_restricted2.compare_f_test(lm_gdppercapita_restricted3)

lm_gdppercapita_restricted4 = smf.ols(formula='gdp_percapita_dividedby1000 ~ try_private + patent_nres_per100people', data=df).fit()
lm_gdppercapita_restricted3.compare_f_test(lm_gdppercapita_restricted4)

lm_gdppercapita = lm_gdppercapita_restricted3
lm_gdppercapita.summary()

COMMENT:

When I perform a f-test on unrestricted and restricted regressions on gdp_percapita_dividedby1000, I see that all random variables except pryntry, try_private and patent_nres_per100people are jointly non-significant. Therefore, I took the restricted regression. 

(Note: I took restricted3 since the last f-test shows that pryntry is significant)

Regressions on gini_index

lm_gini_unrestricted = smf.ols(formula='gini_index ~ pisa_read + pisa_math + pisa_science + pryntry + try_public + try_private', data=df).fit()
lm_gini_unrestricted.summary()

lm_gini_restricted = smf.ols(formula='gini_index ~ pisa_read + pisa_math', data=df).fit()
lm_gini_unrestricted.compare_f_test(lm_gini_restricted)

lm_gini = lm_gini_restricted
lm_gini.summary()

COMMENT:

When I perform a f-test on unrestricted and restricted regressions on gini_index, I see that all random variables except pisa_read and pisa_math are jointly non-significant. Therefore, I took the restricted regression.

ALL REGRESSIONS AND RESULTS:

- pryntry on pisa_read: non-significant
- pryntry on pisa_math: non-significant
- pryntry on pisa_science: non-significant
- pisa_math and try_public on rd_exp: significat, R-squared = 0.543
- try_public on pisa_researchers_rd: significant, R-squared = 0.342
- pisa_read, pisa_math, pisa_science, pryntry, try_public, try_private, rd_exp, researchers_rd on high_tech_exports: non-significant
- try_private, rd_exp on patent_res_per100people: significant, R-squared = 0.593
- try_private, rd_exp on patent_nres_per100people: significant, R-squared = 0.632
- pryntry, try_private, patent_nres_per100people on gdp_percaoita_dividedby1000: significant, R-squared = 0.577
- pisa_read, pisa_math on gini_index: significant, R-squared = 0.725
