==> code/10-CategorizeData.ipynb <==
Import pandas

import pandas as pd

Read csv

df = pd.read_csv("../data/student/student-mat.csv", delimiter = ";")

Bin age, absences and test scores

for i in range(len(df)):
    if (df.get_value(i, 'age') <= 18): 
        df.set_value(i, 'age', 0)
    else:                              
        df.set_value(i, 'age', 1)

for i in range(len(df)):     
    if   (df.get_value(i, 'absences') < 20):
        df.set_value(i, 'absences', 0)
    elif (df.get_value(i, 'absences') < 40):
        df.set_value(i, 'absences', 1)
    elif (df.get_value(i, 'absences') < 60):
        df.set_value(i, 'absences', 2)
    elif (df.get_value(i, 'absences') < 80):
        df.set_value(i, 'absences', 3)
    else:
        df.set_value(i, 'absences', 4)

for i in range(len(df)):     
    if   (df.get_value(i, 'G1') < 4):
        df.set_value(i, 'G1', 0)
    elif (df.get_value(i, 'G1') < 8):
        df.set_value(i, 'G1', 1)
    elif (df.get_value(i, 'G1') < 12):
        df.set_value(i, 'G1', 2)
    elif (df.get_value(i, 'G1') < 16):
        df.set_value(i, 'G1', 3)
    else:
        df.set_value(i, 'G1', 4)
        
for i in range(len(df)):     
    if   (df.get_value(i, 'G2') < 4):
        df.set_value(i, 'G2', 0)
    elif (df.get_value(i, 'G2') < 8):
        df.set_value(i, 'G2', 1)
    elif (df.get_value(i, 'G2') < 12):
        df.set_value(i, 'G2', 2)
    elif (df.get_value(i, 'G2') < 16):
        df.set_value(i, 'G2', 3)
    else:
        df.set_value(i, 'G2', 4)        
        
for i in range(len(df)):     
    if   (df.get_value(i, 'G3') < 4):
        df.set_value(i, 'G3', 0)
    elif (df.get_value(i, 'G3') < 8):
        df.set_value(i, 'G3', 1)
    elif (df.get_value(i, 'G3') < 12):
        df.set_value(i, 'G3', 2)
    elif (df.get_value(i, 'G3') < 16):
        df.set_value(i, 'G3', 3)
    else:
        df.set_value(i, 'G3', 4)

df.to_csv("../data/student/student-mat-categorized.csv", index = False)

df = pd.read_csv("../data/student/student-por.csv", delimiter = ";")

for i in range(len(df)):
    if (df.get_value(i, 'age') <= 18): 
        df.set_value(i, 'age', 0)
    else:                              
        df.set_value(i, 'age', 1)

for i in range(len(df)):     
    if   (df.get_value(i, 'absences') < 20):
        df.set_value(i, 'absences', 0)
    elif (df.get_value(i, 'absences') < 40):
        df.set_value(i, 'absences', 1)
    elif (df.get_value(i, 'absences') < 60):
        df.set_value(i, 'absences', 2)
    elif (df.get_value(i, 'absences') < 80):
        df.set_value(i, 'absences', 3)
    else:
        df.set_value(i, 'absences', 4)

for i in range(len(df)):     
    if   (df.get_value(i, 'G1') < 4):
        df.set_value(i, 'G1', 0)
    elif (df.get_value(i, 'G1') < 8):
        df.set_value(i, 'G1', 1)
    elif (df.get_value(i, 'G1') < 12):
        df.set_value(i, 'G1', 2)
    elif (df.get_value(i, 'G1') < 16):
        df.set_value(i, 'G1', 3)
    else:
        df.set_value(i, 'G1', 4)
        
for i in range(len(df)):     
    if   (df.get_value(i, 'G2') < 4):
        df.set_value(i, 'G2', 0)
    elif (df.get_value(i, 'G2') < 8):
        df.set_value(i, 'G2', 1)
    elif (df.get_value(i, 'G2') < 12):
        df.set_value(i, 'G2', 2)
    elif (df.get_value(i, 'G2') < 16):
        df.set_value(i, 'G2', 3)
    else:
        df.set_value(i, 'G2', 4)        
        
for i in range(len(df)):     
    if   (df.get_value(i, 'G3') < 4):
        df.set_value(i, 'G3', 0)
    elif (df.get_value(i, 'G3') < 8):
        df.set_value(i, 'G3', 1)
    elif (df.get_value(i, 'G3') < 12):
        df.set_value(i, 'G3', 2)
    elif (df.get_value(i, 'G3') < 16):
        df.set_value(i, 'G3', 3)
    else:
        df.set_value(i, 'G3', 4)

Write csv

df.to_csv("../data/student/student-por-categorized.csv", index = False)
