==> code/13-MatZeroR.ipynb <==
Import libraries

import numpy as np
import pandas as pd

Read csv file

df = pd.read_csv("../data/student/student-mat-categorized.csv")

header = list(df)

Convert text to dummy variables

original = []
for attr in header:
    counts = df[attr].value_counts()
    original.append(counts.index.tolist())
    
    for i in range(len(df)):
        value = df.get_value(i, attr)
        new = counts.index.tolist().index(value)
        df.set_value(i, attr, new)

y Values

y = pd.DataFrame.copy(df.iloc[:, -1])

max(y.value_counts())/(len(y))
