==> code/21-Percentages.ipynb <==
Import libraries

import pandas as pd
import numpy as np

Read data

df = pd.read_csv("../data/student/student-mat.csv", delimiter = ";")

for i in range(len(df)):
    if (df.get_value(i, 'age') <= 18): 
        df.set_value(i, 'age', 0)
    else:                              
        df.set_value(i, 'age', 1)

Compute percentages

df_sorted = df.sort_values(['G3'])
df_sorted.index = np.arange(len(df))

percentile = pd.DataFrame.copy(df_sorted)

for i in range(len(df_sorted)):
    percent = i / len(df_sorted)
    
    if percent >= 0.9:
        percentile.set_value(i, "Top10p", "Yes")
    else:
        percentile.set_value(i, "Top10p", "No")
        
    if percent <= 0.1:
        percentile.set_value(i, "Bottom10p", "Yes")
    else:
        percentile.set_value(i, "Bottom10p", "No") 
        
    if percent >= 0.8:
        percentile.set_value(i, "Top20p", "Yes")
    else:
        percentile.set_value(i, "Top20p", "No")
        
    if percent <= 0.2:
        percentile.set_value(i, "Bottom20p", "Yes")
    else:
        percentile.set_value(i, "Bottom20p", "No")
        
df_sorted = pd.concat([df_sorted.iloc[:, :-3], percentile.iloc[:, -4:]], axis=1)

Write data

Read data

df_sorted.to_csv("../data/student/student-mat-percentiled.csv", index = False)

df = pd.read_csv("../data/student/student-por.csv", delimiter = ";")

for i in range(len(df)):
    if (df.get_value(i, 'age') <= 18): 
        df.set_value(i, 'age', 0)
    else:                              
        df.set_value(i, 'age', 1)

Compute percentages

df_sorted = df.sort_values(['G3'])
df_sorted.index = np.arange(len(df))

percentile = pd.DataFrame.copy(df_sorted)

for i in range(len(df_sorted)):
    percent = i / len(df_sorted)
    
    if percent >= 0.9:
        percentile.set_value(i, "Top10p", "Yes")
    else:
        percentile.set_value(i, "Top10p", "No")
        
    if percent <= 0.1:
        percentile.set_value(i, "Bottom10p", "Yes")
    else:
        percentile.set_value(i, "Bottom10p", "No") 
        
    if percent >= 0.8:
        percentile.set_value(i, "Top20p", "Yes")
    else:
        percentile.set_value(i, "Top20p", "No")
        
    if percent <= 0.2:
        percentile.set_value(i, "Bottom20p", "Yes")
    else:
        percentile.set_value(i, "Bottom20p", "No")
        
df_sorted = pd.concat([df_sorted.iloc[:, :-3], percentile.iloc[:, -4:]], axis=1)

Write data

df_sorted.to_csv("../data/student/student-por-percentiled.csv", index = False)
